+++ tests/unit/test_utils.py
@@ -1,7 +1,12 @@
 """Module to test the utils module."""
+from datetime import timedelta
 from unittest import TestCase
 
-from napps.kytos.flow_manager.utils import _valid_consistency_ignored, _validate_range
+from napps.kytos.flow_manager.utils import (
+    _valid_consistency_ignored,
+    _validate_range,
+    get_min_wait_diff,
+)
 
 
 class TestUtils(TestCase):
@@ -26,3 +31,27 @@ class TestUtils(TestCase):
         for values in test_data:
             with self.subTest(values=values):
                 assert not _valid_consistency_ignored(values)
+
+    def test_get_min_wait_diff_early_return(self):
+        """Test get_min_wait diff early return."""
+        test_data = [
+            (timedelta(seconds=1), timedelta(seconds=2), 3),
+            (timedelta(seconds=4), timedelta(seconds=3), 0),
+            (timedelta(seconds=8), timedelta(seconds=2), 4),
+        ]
+        for dt_t2, dt_t1, min_wait in test_data:
+            with self.subTest(dt_t2=dt_t2, dt_t1=dt_t1, min_wait=min_wait):
+                assert get_min_wait_diff(dt_t2, dt_t1, min_wait) == 0
+
+    def test_get_min_wait_diff(self):
+        """Test get_min_wait diff values."""
+        test_data = [
+            (timedelta(seconds=3), timedelta(seconds=2), 4),
+            (timedelta(seconds=5), timedelta(seconds=2), 6),
+        ]
+        for dt_t2, dt_t1, min_wait in test_data:
+            with self.subTest(dt_t2=dt_t2, dt_t1=dt_t1, min_wait=min_wait):
+                assert (
+                    get_min_wait_diff(dt_t2, dt_t1, min_wait)
+                    == min_wait - (dt_t2 - dt_t1).total_seconds()
+                )